# swagger_server/models/artist_metrics.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ArtistMetrics(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, playbacks: int=None, songs: int=None, popularity: int=None):  # noqa: E501
        """ArtistMetrics - a model defined in Swagger

        :param id: The id of this ArtistMetrics.  # noqa: E501
        :type id: int
        :param playbacks: The playbacks of this ArtistMetrics.  # noqa: E501
        :type playbacks: int
        :param songs: The songs of this ArtistMetrics.  # noqa: E501
        :type songs: int
        :param popularity: The popularity of this ArtistMetrics.  # noqa: E501
        :type popularity: int
        """
        self.swagger_types = {
            'id': int,
            'playbacks': int,
            'songs': int,
            'popularity': int
        }

        self.attribute_map = {
            'id': 'id',
            'playbacks': 'playbacks',
            'songs': 'songs',
            'popularity': 'popularity'
        }
        self._id = id
        self._playbacks = playbacks
        self._songs = songs
        self._popularity = popularity

    @classmethod
    def from_dict(cls, dikt) -> 'ArtistMetrics':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ArtistMetrics of this ArtistMetrics.  # noqa: E501
        :rtype: ArtistMetrics
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this ArtistMetrics.

        Unique identifier of the artist  # noqa: E501

        :return: The id of this ArtistMetrics.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this ArtistMetrics.

        Unique identifier of the artist  # noqa: E501

        :param id: The id of this ArtistMetrics.
        :type id: int
        """

        self._id = id

    @property
    def playbacks(self) -> int:
        """Gets the playbacks of this ArtistMetrics.

        Total number of playbacks of the artist  # noqa: E501

        :return: The playbacks of this ArtistMetrics.
        :rtype: int
        """
        return self._playbacks

    @playbacks.setter
    def playbacks(self, playbacks: int):
        """Sets the playbacks of this ArtistMetrics.

        Total number of playbacks of the artist  # noqa: E501

        :param playbacks: The playbacks of this ArtistMetrics.
        :type playbacks: int
        """

        self._playbacks = playbacks

    @property
    def songs(self) -> int:
        """Gets the songs of this ArtistMetrics.

        Total number of songs of the artist  # noqa: E501

        :return: The songs of this ArtistMetrics.
        :rtype: int
        """
        return self._songs

    @songs.setter
    def songs(self, songs: int):
        """Sets the songs of this ArtistMetrics.

        Total number of songs of the artist  # noqa: E501

        :param songs: The songs of this ArtistMetrics.
        :type songs: int
        """

        self._songs = songs

    @property
    def popularity(self) -> int:
        """Gets the popularity of this ArtistMetrics.

        Position of the artist in the popularity ranking  # noqa: E501

        :return: The popularity of this ArtistMetrics.
        :rtype: int
        """
        return self._popularity

    @popularity.setter
    def popularity(self, popularity: int):
        """Sets the popularity of this ArtistMetrics.

        Position of the artist in the popularity ranking  # noqa: E501

        :param popularity: The popularity of this ArtistMetrics.
        :type popularity: int
        """

        self._popularity = popularity